# src/swagger_files/QuickCode.DemoUzeyir.IdentityModule.Api-Swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "QuickCode - DemoUzeyir - IdentityModule Api (1.0.0.0)",
    "termsOfService": "https://www.quickcode.net",
    "contact": {
      "name": "Üzeyir Apaydın",
      "url": "https://www.quickcode.net"
    },
    "license": {
      "name": "QuickCode.Net License",
      "url": "https://www.quickcode.net"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/identity-module/portal-menus/count": {
      "get": {
        "tags": [
          "PortalMenus"
        ],
        "operationId": "PortalMenusCount",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/identity-module/portal-menus/{key}": {
      "get": {
        "tags": [
          "PortalMenus"
        ],
        "operationId": "PortalMenusGetItem",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuickCode.DemoUzeyir.IdentityModule.Application.Dtos.PortalMenu.PortalMenuDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "PortalMenus"
        ],
        "operationId": "PortalMenusUpdate",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuickCode.DemoUzeyir.IdentityModule.Application.Dtos.PortalMenu.PortalMenuDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/QuickCode.DemoUzeyir.IdentityModule.Application.Dtos.PortalMenu.PortalMenuDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/QuickCode.DemoUzeyir.IdentityModule.Application.Dtos.PortalMenu.PortalMenuDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "PortalMenus"
        ],
        "operationId": "PortalMenusDelete",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/identity-module/portal-menus/delete-portal-menu-items-with-module-name/{portalMenuKey}": {
      "delete": {
        "tags": [
          "PortalMenus"
        ],
        "operationId": "PortalMenusDeletePortalMenuItemsWithModuleName",
        "parameters": [
          {
            "name": "portalMenuKey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/identity-module/portal-menus/delete-portal-menu-items-with-model-name/{portalMenuKey}/{portalMenuName}": {
      "delete": {
        "tags": [
          "PortalMenus"
        ],
        "operationId": "PortalMenusDeletePortalMenuItemsWithModelName",
        "parameters": [
          {
            "name": "portalMenuKey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "portalMenuName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/identity-module/portal-page-definitions/get-portal-page-definitions/{permissionGroupName}": {
      "get": {
        "tags": [
          "PortalPageDefinitions"
        ],
        "operationId": "PortalPageDefinitionsGetPortalPageDefinitions",
        "parameters": [
          {
            "name": "permissionGroupName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuickCode.DemoUzeyir.IdentityModule.Application.Models.PortalPageAccessGrantList"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/identity-module/portal-page-definitions/update-portal-page-permission": {
      "post": {
        "tags": [
          "PortalPageDefinitions"
        ],
        "operationId": "PortalPageDefinitionsUpdatePortalPagePermission",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuickCode.DemoUzeyir.IdentityModule.Application.Models.UpdatePortalPageAccessGrantRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/QuickCode.DemoUzeyir.IdentityModule.Application.Models.UpdatePortalPageAccessGrantRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/QuickCode.DemoUzeyir.IdentityModule.Application.Models.UpdatePortalPageAccessGrantRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "QuickCode.DemoUzeyir.IdentityModule.Application.Dtos.PortalMenu.PortalMenuDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "text": {
            "type": "string",
            "nullable": true
          },
          "tooltip": {
            "type": "string",
            "nullable": true
          },
          "actionName": {
            "type": "string",
            "nullable": true
          },
          "orderNo": {
            "type": "integer",
            "format": "int32"
          },
          "parentName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "QuickCode.DemoUzeyir.IdentityModule.Application.Models.PortalPageAccessGrantList": {
        "type": "object",
        "properties": {
          "permissionGroupName": {
            "type": "string",
            "nullable": true
          },
          "portalPageDefinitions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuickCode.DemoUzeyir.IdentityModule.Application.Models.PortalPagePermissionItem"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "QuickCode.DemoUzeyir.IdentityModule.Application.Models.PortalPagePermissionItem": {
        "type": "object",
        "properties": {
          "portalPagePermissionName": {
            "type": "string",
            "nullable": true
          },
          "modelName": {
            "type": "string",
            "nullable": true
          },
          "moduleName": {
            "type": "string",
            "nullable": true
          },
          "portalPagePermissionTypes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuickCode.DemoUzeyir.IdentityModule.Application.Models.PortalPagePermissionTypeItem"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "QuickCode.DemoUzeyir.IdentityModule.Application.Models.PortalPagePermissionTypeItem": {
        "type": "object",
        "properties": {
          "portalPagePermissionType": {
            "$ref": "#/components/schemas/QuickCode.DemoUzeyir.IdentityModule.Domain.Enums.PageActionType"
          },
          "value": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "QuickCode.DemoUzeyir.IdentityModule.Application.Models.UpdatePortalPageAccessGrantRequest": {
        "type": "object",
        "properties": {
          "permissionGroupName": {
            "type": "string",
            "nullable": true
          },
          "portalPagePermissionName": {
            "type": "string",
            "nullable": true
          },
          "portalPagePermissionType": {
            "$ref": "#/components/schemas/QuickCode.DemoUzeyir.IdentityModule.Domain.Enums.PageActionType"
          },
          "value": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "QuickCode.DemoUzeyir.IdentityModule.Domain.Enums.PageActionType": {
        "enum": [
          "List",
          "Detail",
          "Insert",
          "Update",
          "Delete"
        ],
        "type": "string",
        "x-enumNames": [
          "List",
          "Detail",
          "Insert",
          "Update",
          "Delete"
        ],
        "x-enum-varnames": [
          "List",
          "Detail",
          "Insert",
          "Update",
          "Delete"
        ],
        "x-enumDescriptions": [
          "List page",
          "Details / view page",
          "Create page",
          "Update page",
          "Delete page"
        ]
      }
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "description": "Please enter a valid Api Key",
        "name": "X-Api-Key",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "ApiKey": [ ]
    }
  ]
}